Develop -> Master

+++ src/routers/cards.js
@@ -1,7 +1,8 @@
 const express = require('express')
-const Cards = require('../models/cards')
+const { Cards } = require('../models/cards')
 const auth = require('../middleware/auth')
 const hbs = require('hbs')
+const Boards = require('../models/boards')
 
 const router = new express.Router()
 
@@ -18,7 +19,7 @@ router.post('/add/card', auth, async (req, res) => {
     try {
         const card = new Cards({username, title, description, status, completed})
         await card.save()
-        res.status(301).redirect('/kanban-board')
+        res.status(301).redirect('/progress-board')
     } catch (e) {
         res.status(400).send({
             error: 'Card cannot be added.'
@@ -28,17 +29,29 @@ router.post('/add/card', auth, async (req, res) => {
 
 // Update card on Kanban Board.
 router.post('/update/card', auth, async (req, res) => {
-    const title = req.body.update_card_title
-    const description = req.body.update_card_description
-    const status = req.body.update_card_status
+    let title = req.body.update_card_title
+    let description = req.body.update_card_description
+    let status = req.body.update_card_status
     const card_id = req.body.card_id
+    const board_id = req.body.board_id
     let completed = false
     if (status === 'done') {
         completed = true
     }
     try {
-        await Cards.findByIdAndUpdate(card_id, {title, description, status, completed})
-        res.status(301).redirect('/kanban-board')
+        if (board_id) {
+            await Boards.findOneAndUpdate({ _id: board_id, 'cards._id': card_id}, {
+                $set: {
+                    'cards.$.title': title,
+                    'cards.$.description': description,
+                    'cards.$.status': status,
+                }
+            })
+            res.status(301).redirect('/board/' + board_id)
+        } else {
+            await Cards.findByIdAndUpdate(card_id, {title, description, status, completed})
+            res.status(301).redirect('/progress-board')
+        }
     } catch (e) {
         res.status(400).send({
             error: 'Card cannot be updated.'
@@ -47,36 +60,50 @@ router.post('/update/card', auth, async (req, res) => {
 })
 
 // Delete card on Kanban Board.
-router.get('/remove/card', auth, async (req, res) => {
-    const card_id = req.query.id
-    try {
+router.delete('/remove/card', auth, async (req, res) => {
+    const card_id = req.body.card_id
+    const board_id = req.body.board_id
+    if (board_id !== '') {
+        try {
+            await Boards.findOneAndUpdate({ _id: board_id }, {
+                $pull: {
+                    'cards': {
+                        '_id': card_id
+                    }
+                }
+            })
+            res.send({
+                card_id,
+                board_id
+            })
+        } catch (e) {
+            res.status(400).send({
+                error: 'Card cannot be deleted.'
+            })
+        }
+    } else {
         await Cards.findByIdAndDelete(card_id)
-        res.status(301).redirect('/kanban-board')
-    } catch (e) {
-        res.status(400).send({
-            error: 'Card cannot be deleted.'
+        res.send({
+            card_id
         })
-    }
-})
-
-// Fetch Kanban Board for a user.
-router.get('/kanban-board', auth, async (req, res) => {
-    const username = req.user.username
-    try {
-        const cards = await Cards.find({ username })
-        res.render('kanban', {
-            cards
-        })
-    } catch (e) {
-        res.render('kanban')
     }
 })
 
 // Fetch card through ajax.
 router.get('/fetch/card', auth, async (req, res) => {
     const card_id = req.query.id
+    const board_id = req.query.board_id
     try {
-        const card = await Cards.findById(card_id)
+        let card
+        if (board_id !== '') {
+            const board = await Boards.findById(board_id)
+            card = board.cards.filter(card => {
+                return card.id == card_id
+            });
+            card = card[0]
+        } else {
+            card = await Cards.findById(card_id)
+        }
         res.send({
             card
         })
@@ -89,23 +116,39 @@ router.get('/fetch/card', auth, async (req, res) => {
 router.get('/update/card', auth, async (req, res) => {
     const card_id = req.query.id
     const status = req.query.status
+    const board_id = req.query.board_id
     let completed = false
     if (status === 'done') {
         completed = true
     }
-    try {
-        const card = await Cards.findByIdAndUpdate(card_id, {status, completed})
-        res.send({
-            card
-        })
-    } catch (e) {
-        res.render('kanban')
+    if (!board_id) {
+        try {
+            const card = await Cards.findByIdAndUpdate(card_id, {status, completed})
+            res.send({
+                card
+            })
+        } catch (e) {
+            res.render('kanban')
+        }
+    } else {
+        try {
+            const board = await Boards.findOneAndUpdate({ _id: board_id, 'cards._id': card_id}, {
+                $set: {
+                    'cards.$.status': status
+                }
+            })
+            res.send({
+                board
+            })
+        } catch (e) {
+            res.render('board')
+        }
     }
 })
 
 // Register helper for handlebars.
 hbs.registerHelper('cardsStatus', function(card_status, status, options) {
-    if (card_status === status) {
+    if (card_status == status) {
         return options.fn(this)
     } else {
         options.inverse(this)
